requirements.txt
# Strands Agents + AgentCore Memory 統合
strands-agents>=0.1.0
bedrock-agentcore[strands-agents]>=1.1.2

# AWS SDK
boto3>=1.35.0

# Web Server (AgentCore Runtime uses uvicorn internally)
uvicorn>=0.30.0
